third_party/blink/web_tests/platform/generic/http/tests/devtools/sources/debugger-pause/pause-in-internal-script-expected.txt

Tests that internal scripts unknown to front-end are processed correctly when appear in debugger call frames. Bug 64995


Running: testSetBreakpoint
Script source was shown.
Set timer for test function.
Script execution paused.
Call stack:
    0)  (pause-in-internal-script.js:22)
    1) testFunction (pause-in-internal-script.js:20)
Script execution resumed.
